zcash_client_sqlite: Remove the remainder of the sqlite dependencies from TestState
